DFTK.jl: The density-functional toolkit.

DFTK.jl: The density-functional toolkit.

DFTK is a julia package of for playing with plane-wave density-functional theory algorithms.

Terminology and Definitions

The general terminology used throughout the documentation of the plane-wave aspects of the code.

Lattices

Usually we denote with $A$ the matrix containing all lattice vectors as columns and with

\[\textbf{B} = 2\pi \textbf{A}^{-T}\]

the matrix containing the reciprocal lattice vectors as columns.

Units

Unless otherwise stated the code and documentation uses atomic units and fractional or integer coordinates for $k$-Points and wave vectors $G$. The equivalent Vectors in cartesian coordiates will be denoted as $k^c$ or $G^c$, i.e.

\[k^c = \textbf{B} k \quad G^c = \textbf{B} G.\]

Plane wave basis functions

At the moment the code works exclusively with orthonormal plane waves. In other words our bases consist of functions

\[e_{G^c} = 1/\sqrt{\Omega} e^{i\, G^c \cdot x}\]

where $\Omega$ is the unit cell volume and $G^c$ is a wave vector in cartesian coordiates.

Basis sets

The wave-function basis $B_k$, consisting of all plane-wave basis functions below the desired energy cutoff $E_\text{cut}$ for each $k$-point:

\[B_k = \{ e_{G^c} : 1/2 |G^c + k^c|^2 ≤ E_\text{cut}.\]

Geometrically the corresponding wave vectors $G^c$ form a ball of radius $\sqrt{2 E_\text{cut}}$ centred at $k^c$. This makes the corresponding set of $G$-vectors

\[\{ G : |\textbf{B} (G + k)| ≤ 2 \sqrt{E_\text{cut}} \}\]

in integer coordinates an ellipsoid.

The potential or density basis $B_\rho$, consisting of all plane waves on which a potential needs to be known in order to be consistent with the union of all $B_k$ for all $k$. This means that it is the set

\[B_\rho = \{ e_{G^c} - e_{\tilde{G}^c} : e_{G^c}, e_{\tilde{G}^c} \in B_k \}.\]

This is equivalent to the alternative definition

\[B_\rho = \{ e_{G^c} : 1/2 |G^c|^2 ≤ α^2 E_\text{cut} \},\]

for a supersampling factor $\alpha = 2$. Geometrically this is again a ball in cartesian coordinates and an ellipsoid in integer coordinates.

In practice we do not use $B_\rho$ in the code, since fast-fourier transforms (FFT) operate on rectangular grids instead. For this reason the code determines $C_\rho$, the smallest rectangular grid in integer coordinates which contains all $G$-vectors corresponding to the plane waves of $B_\rho$. For this we take

\[C_\rho = \{ G = (G_1, G_2, G_3)^T : |G_i| ≤ N_i \}\]

where the bounds $N_i$ are determined as follows. Since $G = \textbf{B}^{-1} G^c$ one can employ Cauchy-Schwartz to get

\[N_i = max_{|G^c|^2 ≤ 2 α^2 E_\text{cut}}(\textbf{B}^{-1}[i, :] \cdot G^c) ≤ |\textbf{B}^{-1}[i, :]| \sqrt{2 α^2 E_\text{cut}}.\]

With $\textbf{B}^{-1} = frac{1}{2\pi} \textbf{A}^T$ therefore

\[N_i ≤ |\textbf{A}[:, i]| \frac{\sqrt{2 α^2 E_\text{cut}}}{2π}\]

where e.g. $\textbf{A}[:, i]$ denotes the $i$-th column of $\textbf{A}$. Notice, that this makes $C_\rho$ is a rectangular shape in integer coordinates, but a parallelepiped in cartesian coordinates.

TODO not yet properly updated from here

The XC basis $B_\text{XC}$, which is used for computing the application of the exchange-correlation potential operator to the density $\rho$, represented in the basis $B_\rho$, that is
\[B_\text{XC} = \{e_G : 1/2 |G|_\text{max}^2 ≤ β E_\text{cut} \}.\]
Since the exchange-correlation potential might involve arbitrary powers of the density $ρ$, a numerically exact computation of the integral
\[\langle e_G | V_\text{XC}(ρ) e_{G'} \rangle \qquad \text{with} \qquad e_G, e_{G'} ∈ B_{Ψ,k}\]
requires the exchange-correlation supersampling factor $\beta$ to be infinite. In practice, $\beta =4$ is usually chosen, such that $B_\text{XC} = B_\rho$.

Real-space grids

Due to the Fourier-duality of reciprocal-space and real-space lattice, the above basis sets define corresponding real-space grids as well:

The grid $B_\rho^\ast$, the potential integration grid, which is the grid used for convolutions of a potential with the discretized representation of a DFT orbital. It is simply the iFFT-dual real-space grid of $B_\rho$.

The grid $B^\ast_\text{XC}$, the exchange-correlation integration grid, i.e. the grid used for convolutions of the exchange-correlation functional terms with the density or derivatives of it. It is the iFFT-dual of $B_\text{XC}$.

Core

DFTK.PlaneWaveBasis — Type.

TODO docme

fft_size is now Fourier grid size kcoords is vector of Vec3

source

Missing docstring.

Missing docstring for set_kpoints!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for basis_ρ. Check Documenter's build log for details.

DFTK.G_to_r! — Function.

G_to_r!(f_real, pw::PlaneWaveBasis, [kpt::Kpoint, ], f_fourier)

Perform an iFFT to translate between f_fourier, a fourier representation of a function either on $B_k$ (if kpt is given) or on $C_ρ$ (if not), and f_real. The function will destroy all data in f_real.

source

DFTK.r_to_G! — Function.

r_to_G!(f_fourier, pw::PlaneWaveBasis, [kpt::Kpoint, ], f_real)

Perform an FFT to translate between f_real, a function represented on $C_ρ^\ast$ and its fourier representation. Truncatate the fourier coefficients to $B_k$ (if kpt is given). Note: If kpt is given, all data in $f_real$ will be distroyed as well.

source

Missing docstring.

Missing docstring for PotLocal. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PotNonLocal. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Kinetic. Check Documenter's build log for details.

DFTK.Hamiltonian — Type.

Initialise a one-particle Hamiltonian from a model and optionally a density.

source

Missing docstring.

Missing docstring for apply_hamiltonian!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DFTK.apply_fourier!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DFTK.apply_real!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DFTK.update_potential!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DFTK.update_energies_potential!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DFTK.empty_potential. Check Documenter's build log for details.

DFTK.PreconditionerKinetic — Type.

Kinetic-energy based preconditioner.

Applies $1 / (|k + G|^2 / 2 + α)$ to the vectors, when called with ldiv!. This attempts to dampen the high-kinetic energy parts of the Hamiltonian, thus making the Hamiltonian more well-conditioned.

source

Missing docstring.

Missing docstring for DFTK.lobpcg. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DFTK.lobpcg_qr. Check Documenter's build log for details.

DFTK.lobpcg_scipy — Function.

lobpcg_scipy(A, X0; prec=nothing, tol=nothing, largest=false, kwargs...)

Call scipy's version of LOBPCG for finding the eigenpairs of a Hermitian matrix A. X0 is the set of guess vectors, also determining the number of eigenpairs to be sought.

source

Missing docstring.

Missing docstring for DFTK.lobpcg_itsolve. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DFTK.occupation_zero_temperature. Check Documenter's build log for details.

Missing docstring.

Missing docstring for self_consistent_field. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DFTK.scf_nlsolve. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DFTK.scf_damped. Check Documenter's build log for details.

DFTK.PspHgh — Type.

PspHgh(Zion::Number, rloc::Number, cloc::Vector, rp::Vector, h::Vector;
       identifier="", description="")

Construct a Hartwigsen, Goedecker, Teter, Hutter separable dual-space Gaussian pseudopotential (1998). The required parameters are the ionic charge Zion (total charge - valence electrons), the range for the local Gaussian charge distribution rloc, the coefficients for the local part cloc, the projector radius rp (one per AM channel) and the non-local coupling coefficients between the projectors h (one matrix per AM channel).

source

DFTK.eval_psp_projection_radial — Function.

eval_psp_projection_radial(psp::PspHgh, i, l, qsq::Number)

Evaluate the radial part of the i-th projector for angular momentum l at the reciprocal lattice point with modulus squared qsq.

p(qsq) = ∫{R+} r^2 p(r) jl(q r) dr

HGH98 except they do it with plane waves normalized by 1/sqrt(Ω).

source

DFTK.eval_psp_local_real — Function.

eval_psp_local_real(psp, r)

Evaluate the local part of the pseudopotential in real space. The vector r should be given in cartesian coordinates.

GTH98

source

DFTK.eval_psp_local_fourier — Function.

eval_psp_local_fourier(psp, ΔG)

Evaluate the local part of the pseudopotential in reciprocal space.

This function computes V(q) = ∫R^3 Vloc(r) e^{-iqr} dr = 4π ∫{R+} sin(qr)/q r e^{-iqr} dr

GTH98 except they do it with plane waves normalized by 1/sqrt(Ω).

source

DFTK.compute_density — Function.

compute_density(pw::PlaneWaveBasis, Psi::AbstractVector, occupation::AbstractVector)

Compute the density for a wave function Psi discretised on the plane-wave grid pw, where the individual k-Points are occupied according to occupation. Psi should be one coefficient matrix per k-Point.

source

Utilities

DFTK.Species — Type.

Species(Znuc; psp=nothing)

Structure collecting information about a species in a structure or crystal.

Znuc: Nuclear charge (single atom) or accumulated nuclear change (composition)

psp: Pseudopotential to be used (or nothing for all-electron treatment)

source

DFTK.determine_grid_size — Function.

determine_grid_size(lattice, Ecut; supersampling=2)

Determine the minimal grid size for the fourier grid $C_ρ$ subject to the kinetic energy cutoff Ecut for the wave function and a density supersampling factor. Optimise the grid afterwards for the FFT procedure by ensuring factorisation into small primes. The function will determine the smallest cube $C_ρ$ containing the basis $B_ρ$, i.e. the wave vectors $|G|^2/2 \leq E_\text{cut} ⋅ \text{supersampling}^2$. For an exact representation of the density resulting from wave functions represented in the basis $B_k = \{G : |G + k|^2/2 \leq E_\text{cut}\}$, supersampling should be at least 2.

source

Missing docstring.

Missing docstring for build_local_potential. Check Documenter's build log for details.

Missing docstring.

Missing docstring for build_nonlocal_projectors. Check Documenter's build log for details.

Missing docstring.

Missing docstring for kblock_as_matrix. Check Documenter's build log for details.

DFTK.load_psp — Function.

load_psp(identifier; datadir_psp)

Load a pseudopotential file from the library of pseudopotentials. The file is searched in the directory datadir_psp and by the identifier.

source

DFTK.guess_density — Function.

guess_density(basis, composition...)

Build a superposition of atomic densities (SAD) guess density. The atoms/species are specified in composition as pairs representing a mapping from Species objects to a list of positions in fractional coordinates.

We take for the guess density a gaussian centered around the atom, of length specified by atom_decay_length, normalized to get the right number of electrons

\[\hat{ρ}(G) = Z \exp\left(-(2π \text{length} |G|)^2\right)\]

source

Missing docstring.

Missing docstring for guess_hcore. Check Documenter's build log for details.

Definition of builder functions

Definition of builder functions
   builder(basis::PlaneWaveBasis, energy::Ref, potential; ρ=nothing, Ψ=nothing, kwargs...)
       energy       nothing or reference
       potential    nothing or Array or data structure

   energy !== nothing
       Reference updates with new energy value
   potential !== nothing
       Potential array update with new potential values
       or data structure to use for updating potential data
   energy !== nothing and potential !== nothing
       compute potential values and energies and return energy, potential
       May reuse the storage of the `potential` object

Note: - It is *not* required nor guaranteed that the returned `potential` is an array.
        E.g. for the non-local pseudopotential term it will be a PotNonLocal datastructure.
        The returned energy value, however, is always a scalar.
      - The arguments `ρ` or `Ψ` are not required for calling the function, but the
        function might throw if they are for the requested computations.

Because we do not explicitly make sure the Fourier representation of an analytic potential is a real fucntion in real space only the real part of the real-space representation of a potential is physically meaningful.