DFTK.jl: The density-functional toolkit.

DFTK.jl: The density-functional toolkit.

DFTK is a julia package of for playing with plane-wave density-functional theory algorithms.

Terminology and Definitions

General terminology used throughout the documentation of the plane-wave aspects of the code.

Plane wave basis functions

At the moment the code works exclusively with orthonormal plane waves. In other words our bases consist of functions

\[e_G = 1/\sqrt{\Omega} e^{i\, G \cdot x}\]

where $\Omega$ is the unit cell volume.

Basis sets

The wave-function basis $B_{Ψ,k}$ (used to be $X_k$), consisting of all plane-wave basis functions below the desired energy cutoff $E_\text{cut}$ for each $k$-point:
\[B_{Ψ,k} = \{ e_G : 1/2 |G + k|^2 ≤ E_\text{cut}\]

The potential or density basis $B_\rho$, consisting of all plane waves on which a potential needs to be known in order to be consistent with the union of all $B_{Ψ,k}$ for all $k$. In practice we do not take the smallest possible set of wave vectors $G$ for this, but instead the smallest cubic grid, which satisfies this, i.e.
\[B_\rho = \{ e_G : 1/2 |G|_\text{max}^2 ≤ α E_\text{cut} \},\]
where a supersampling factor $\alpha = 4$ is required to give a numerically exact result, since
\[B_\rho = \{ e_{G+G'} : ∃k e_G, e_{G'} ∈ B_{Ψ,k} \}.\]
The choice of using a cubic grid is done in order to be consistent with usual fast Fourier transform implementations, which work on cubic Fourier grids.

The XC basis $B_\text{XC}$, which is used for computing the application of the exchange-correlation potential operator to the density $\rho$, represented in the basis $B_\rho$, that is
\[B_\text{XC} = \{e_G : 1/2 |G|_\text{max}^2 ≤ β E_\text{cut} \}.\]
Since the exchange-correlation potential might involve arbitrary powers of the density $ρ$, a numerically exact computation of the integral
\[\langle e_G | V_\text{XC}(ρ) e_{G'} \rangle \qquad \text{with} \qquad e_G, e_{G'} ∈ B_{Ψ,k}\]
requires the exchange-correlation supersampling factor $\beta$ to be infinite. In practice, $\beta =4$ is usually chosen, such that $B_\text{XC} = B_\rho$.

Real-space grids

Due to the Fourier-duality of reciprocal-space and real-space lattice, the above basis sets define corresponding real-space grids as well:

The grid $B_\rho^\ast$, the potential integration grid, which is the grid used for convolutions of a potential with the discretized representation of a DFT orbital. It is simply the iFFT-dual real-space grid of $B_\rho$.

The grid $B^\ast_\text{XC}$, the exchange-correlation integration grid, i.e. the grid used for convolutions of the exchange-correlation functional terms with the density or derivatives of it. It is the iFFT-dual of $B_\text{XC}$.

Core

DFTK.PlaneWaveBasis — Type.

PlaneWaveBasis(lattice::Mat3{T}, grid_size::Vec3{I}, Ecut::Number, kpoints,
               kweights) where {T <: Real, I <: Integer}

Create a plane-wave basis from a specification for the Fourier grid size and a kinetic energy cutoff to select the $k$-point-specific wave function basis $B_{Ψ,k}$ in a way that the selected $G$ wave vectors satisfy $|G + k|^2/2 \leq Ecut$.

Examples

julia> b = PlaneWaveBasis(TODO)

Arguments

lattice: Real-space lattice vectors in columns

grid_size: Size of the rectangular Fourier grid used as the density basis $B_ρ$. In each dimension idim the range of wave vectors (in integer coordinates) extends from -ceil(Int, (grid_size[idim]-1) / 2) up to floor(Int, (grid_size[idim]-1) / 2). No optimisation is done on the size of the grid with respect to performing FFTs.

Ecut: Kinetic energy cutoff in Hartree

kpoints: List of $k$-Points in fractional coordinats

kweights: List of corresponding weights for the Brillouin-zone integration.

source

DFTK.set_kpoints! — Function.

Reset the kpoints of an existing Plane-wave basis and change the basis accordingly.

source

DFTK.basis_ρ — Function.

Return a generator producing the range of wave-vector coordinates contained in the Fourier grid $B_ρ$ described by the plane-wave basis.

source

DFTK.G_to_r! — Function.

G_to_r!(pw::PlaneWaveBasis, f_fourier, f_real[, basis_ρ])

Perform an in-place FFT to translate between f_fourier, a fourier representation of a function using the wave vectors specified in basis_ρ and a representation on the real-space density grid $B^∗_ρ$. The function will destroy all data in f_real. If basis_ρ is absent, the full density grid $B_ρ$ is used.

source

DFTK.r_to_G! — Function.

r_to_G!(pw::PlaneWaveBasis, f_real, f_fourier[, gcoords])

Perform an in-place FFT to translate between f_real, a representation of a function on the real-space density grid $B^∗_ρ$ and a fourier representation using the wave vectors specified in gcoords. If gcoords is less than the wave vectors required to exactly represent f_real, than this function implies a truncation. On call all data in f_real and f_fourier will be destroyed. If gcoords is absent, the full density grid $B_ρ$ is used.

source

DFTK.PotLocal — Type.

Class for holding the values of a local potential, like the local part of a pseudopotential

source

DFTK.PotNonLocal — Type.

PotNonLocal(pw::PlaneWaveBasis, positions, psps)

Build a Kleinman-Bylander representation of the non-local potential term for the given basis. positions is a mapping from an identifier to a set of positions in fractional coordinates and psps is a mapping from the identifier to the pseudopotential object associated to this idendifier.

source

DFTK.Kinetic — Type.

Kinetic energy operator in a plane-wave basis.

source

DFTK.Hamiltonian — Type.

Hamiltonian(basis::PlaneWaveBasis; kinetic=Kinetic(basis), pot_local=nothing,
            pot_nonlocal=nothing, pot_hartree=nothing, pot_xc=nothing)

Hamiltonian discretized in a plane-wave basis. Terms which are not specified (left as nothing) are ignored during application. If only a basis object is specified, a free-electron Hamiltonian is constructed. The kinetic pot_local and pot_nonlocal terms shall not contain a non-linearity in the density.

source

DFTK.apply_hamiltonian! — Function.

apply_hamiltonian!(out, ham, ik, pot_hartree_values, pot_xc_values, in)

Apply the ik-th $k$-Point block of a Hamiltonian using precomputed values for the non-linear Hartree and XC potential (in pot_hartree_values and pot_xc_values) on the real-space grid $B_\rho^∗$ at the current density.

For this kinetic and pot_nonlocal are applied to in using apply_fourier!. The other terms are treated via a convolution on the density grid $B_\rho^∗$, that is the procedure

 X_k --pad--> Y --IFFT--> Y* --V--> Y* --FFT--> Y --trunc--> X_k

where

  pad      Zero padding
  FFT      fast-Fourier transform
  IFFT     inverse fast-Fourier transform
  trunc    Truncation to a smaller basis
  V        Apply potential elementwise using `apply_real!`

source

DFTK.apply_fourier! — Function.

Apply a $k$-block of a Hamiltonian term in Fourier space

source

DFTK.apply_real! — Function.

Apply a Hamiltonian term by computation on the real-space density grid $B^∗_ρ$

source

DFTK.compute_potential! — Function.

Compute the potential of a non-linear term (e.g. pot_hartree or pot_xc) on the real-space density grid $B^∗_ρ$, given a current density ρ in the density basis $B_ρ$. If the passed term is nothing, nothing is returned by the function as well, else an array of values.

source

DFTK.empty_potential — Function.

Return an appropriately sized container for a potential term on the real-space grid $B^∗_ρ$

source

DFTK.PreconditionerKinetic — Type.

Kinetic-energy based preconditioner.

Applies $1 / (|k + G|^2 / 2 + α)$ to the vectors, when called with ldiv!. This attempts to dampen the high-kinetic energy parts of the Hamiltonian, thus making the Hamiltonian more well-conditioned.

source

IterativeSolvers.lobpcg — Function.

lobpcg(ham::Hamiltonian, pot_hartree_values, pot_xc_values, nev_per_kpoint::Int;
       guess=nothing, prec=nothing, tol=1e-6, maxiter=100, backend=:lobpcg_qr,
       kwargs...)

Run the LOBPCG implementation from backend for each $k$-Point of the Hamiltonian ham, solving for the nev_per_kpoint smallest eigenvalues. Optionally a guess and a preconditioner prec can be used. pot_hartree_values and pot_xc_values are the precomputed values of the Hartee and XC term of the Hamiltonian on the grid $B_ρ^∗$. The backend parameters selects the LOBPCG implementation to use.

source

DFTK.lobpcg_qr — Function.

lobpcg_qr(A, X0; maxiter=100, prec=I, tol=20size(A,2)*eps(eltype(A)), largest=false)

Naive LOBPCG for finding the largest eigenpairs of a Hermitian matrix A starting from a guess X0 of as many guess vectors as eigenpairs are sought. Optionally a preconditioner prec may be employed.

source

DFTK.lobpcg_scipy — Function.

lobpcg_scipy(A, X0; prec=nothing, tol=nothing, largest=false, kwargs...)

Call scipy's version of LOBPCG for finding the eigenpairs of a Hermitian matrix A. X0 is the set of guess vectors, also determining the number of eigenpairs to be sought.

source

DFTK.lobpcg_itsolve — Function.

lobpcg_itsolve(A, X0; prec=nothing, largest=false, kwargs...)

Call the lobpcg version from the IterativeSolvers package passing through most arguments

source

DFTK.occupation_zero_temperature — Function.

Compute the occupation at zero temperature and without smearing for n_elec electrons and the bands Psi with associated energies.

source

DFTK.self_consistent_field — Function.

self_consistent_field(ham::Hamiltonian, n_bands::Int, n_electrons::Int;
                      ρ=nothing, tol=1e-6, max_iter=100, algorithm=:scf_nlsolve,
                      lobpcg_prec=PreconditionerKinetic(ham, α=0.1))

Run a self-consistent field iteration for the Hamiltonian ham, returning the self-consistnet density, Hartree potential values and XC potential values. n_bands selects the number of bands to be computed, n_electrons the number of electrons, ρ is the initial density, e.g. constructed via a SAD guess. lobpcg_prec specifies the preconditioner used in the LOBPCG algorithms used for diagonalisation. Possible algorithms are :scf_nlsolve or :scf_damped.

source

DFTK.scf_nlsolve — Function.

TODO docme

source

DFTK.scf_damped — Function.

TODO docme

source

DFTK.PspHgh — Type.

PspHgh(Zion::Number, rloc::Number, cloc::Vector, rp::Vector, h::Vector;
       identifier="", description="")

Construct a Hartwigsen, Goedecker, Teter, Hutter separable dual-space Gaussian pseudopotential (1998). The required parameters are the ionic charge Zion (total charge - valence electrons), the range for the local Gaussian charge distribution rloc, the coefficients for the local part cloc, the projector radius rp (one per AM channel) and the non-local coupling coefficients between the projectors h (one matrix per AM channel).

source

DFTK.eval_psp_projection_radial — Function.

eval_psp_projection_radial(psp::PspHgh, i, l, qsq::Number)

Evaluate the radial part of the i-th projector for angular momentum l at the reciprocal lattice point with modulus squared qsq. Compared to the expressions in the GTH and HGH papers, this expression misses a factor of 1/sqrt(Ω).

source

DFTK.eval_psp_local_real — Function.

eval_psp_local_real(psp, r)

Evaluate the local part of the pseudopotential in real space. The vector r should be given in cartesian coordinates.

source

DFTK.eval_psp_local_fourier — Function.

eval_psp_local_fourier

Evaluate the local part of the pseudopotential in reciprocal space. Computes <eG|Vloc|e{G+ΔG}> without taking into account the structure factor and the (4π / Ω) spherical Hankel transform prefactor. ΔG should be in cartesian coordinates.

source

DFTK.compute_density — Function.

compute_density(pw::PlaneWaveBasis, Psi::AbstractVector, occupation::AbstractVector;
                tolerance_orthonormality)

Compute the density for a wave function Psi discretised on the plane-wave grid pw, where the individual k-Points are occupied according to occupation. Psi should be one coefficient matrix per k-Point. If tolerance_orthonormality is ≥ 0, some orthonormality properties are verified explicitly.

source

Utilities

DFTK.determine_grid_size — Function.

determine_grid_size(lattice, Ecut; kpoints=[[0,0,0]], supersampling=2)

Determine the minimal grid size for the density fourier grid $B_ρ$ subject to the kinetic energy cutoff Ecut for the wave function and a density supersampling factor.

The function will determine the union of wave vectors $G$ required to satisfy $|G + k|^2/2 \leq E_\text{cut} ⋅ \text{supersampling}^2$ for all $k$-Points. The returned grid dimensions are the smallest cartesian box to incorporate these $G$.

For an exact representation of the density resulting from wave functions represented in the basis $B_ρ = \{G : |G + k|^2/2 \leq Ecut\}$, supersampling should be at least 2.

source

DFTK.build_local_potential — Function.

build_local_potential(pw::PlaneWaveBasis, positions, generators;
                      compensating_background=true)

Function generating a local potential on the real-space density grid $B^∗_ρ$ defined by the plane-wave basis pw. The potential is generated by summing (in Fourier space) analytic contributions from all species involved in the lattice, followed by an iFFT. The lattice definition is taken implicitly from pw.

The contributions are defined by the two maps positions and generators. The former maps a set of keys, uniquely identifying the lattice species, to a list of fractional coordinates defining their real-space positions. The latter maps each identifier to a function G -> potential(G), which defines the potential value for this species at this reciprocal space position (also passed to this function in integer coordinates). The parameter compensating_background (default true) determines whether the DC component will be automatically set to zero, which physically corresponds to including a compensating change background in the potential model.

Examples

Given an appropriate lattice and basis definition in basis one may build the local potential for an all-electron treatment of sodium chloride as such

julia> build_local_potential(basis,
                             [:Na => [[0,0,0], [1/2,1/2,0], [1/2,0,1/2], [0,1/2,1/2]],
                              :Cl => [[0,1/2,0], [1/2,0,0], [0,0,1/2], [1/2,1/2,1/2]]],
                             [:Na => -11 / sum(abs2, basis.recip_lattice * G), :Cl => -17 / sum(abs2, recip_lattice * G)])

since sodium has nuclear charge 11 and chlorine charge 17.

For crystals composed of only a single species, such as silicon, one can drop the mappings:

julia> build_local_potential(basis, [[0,0,0], [1/8, 1/8, 1/8]], G -> -12 / sum(abs2, basis.recip_lattice * G))

source

DFTK.kblock_as_matrix — Function.

Compute a $k$-Point block of an operator as a dense matrix

source

DFTK.load_psp — Function.

load_psp(identifier; search_directory)

Load a pseudopotential file from the library of pseudopotentials. The file is searched in the directory search_directory and by the identifier.

source

DFTK.guess_gaussian_sad — Function.

guess_gaussian_sad(basis, positions, Znucls, [Zions])

Build a superposition of atomic densities (SAD) guess density. The atoms/species are specified by the three dictionaries positions, Znucls and Zions. The first specifies the list of atom positions, Znucls is the corresponding nuclear charge and Zions the corresponding ionic charge (i.e. the charge left over when pseudopotentials have been taken into account).

source

DFTK.guess_hcore — Function.

Obtain a guess density form diagonalising the core Hamiltonian associated to ham, i.e. strip the non-linear parts. compute_occupation computes the occupation values, taking into account potential smearing. lobpcg_prec is a preconditioner for LOBPCG and n_bands the number of bands to compute.

source